+++ src/main/java/module-info.java
@@ -1,4 +1,4 @@
-module gu {
+module pihole {
     requires javafx.controls;
     requires javafx.fxml;
     requires javafx.web;
@@ -9,9 +9,9 @@ module gu {
     requires org.kordamp.ikonli.javafx;
     requires org.kordamp.bootstrapfx.core;
     requires eu.hansolo.tilesfx;
-    requires json.simple;
     requires org.jetbrains.annotations;
+    requires json.simple;
 
-    opens gu to javafx.fxml;
-    exports gu;
+    opens pihole to javafx.fxml;
+    exports pihole;
 }